chore: Update e2e test for PageSpeed Insights link and package dependencies
